requirements.txt
# Django
Django>=4.2.0,<5.3.0
django-allauth>=0.54.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.0
django-taggit>=4.0.0
django-markdownx>=4.0.0
django-ckeditor>=6.0.0
django-debug-toolbar>=4.0.0
Pillow>=10.0.0
whitenoise>=6.0.0

# Database
psycopg2-binary>=2.9.0

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0

# Production
gunicorn>=21.0.0
python-dotenv>=1.0.0

# Testing
coverage>=7.0.0
